# memox-api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "DeckResponse" : {
        "properties" : {
          "contentType" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "parentDeckId" : {
            "type" : "string"
          },
          "rootDeckId" : {
            "type" : "string"
          },
          "schedulerGeneration" : {
            "format" : "int32",
            "type" : "integer"
          },
          "schedulerType" : {
            "type" : "string"
          },
          "schedulerVersion" : {
            "format" : "int32",
            "type" : "integer"
          },
          "siblingPosition" : {
            "format" : "int32",
            "type" : "integer"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "DeleteBatchResponse" : {
        "properties" : {
          "deletedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "itemType" : {
            "type" : "string"
          },
          "rootItemId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "DeleteCardsRequest" : {
        "properties" : {
          "cardIds" : {
            "items" : {
              "type" : "string"
            },
            "minItems" : 1,
            "type" : "array"
          }
        },
        "required" : [
          "cardIds"
        ],
        "type" : "object"
      },
...
      "PurgeBatchesRequest" : {
        "properties" : {
          "batchIds" : {
            "items" : {
              "type" : "string"
            },
            "minItems" : 1,
            "type" : "array"
          }
        },
        "required" : [
          "batchIds"
        ],
        "type" : "object"
      },
      "PurgeReportResponse" : {
        "properties" : {
          "purgedBatches" : {
            "format" : "int32",
            "type" : "integer"
          },
          "skippedBatches" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "RenameDeckRequest" : {
        "properties" : {
          "name" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [
          "name"
        ],
        "type" : "object"
      },
...
      "RestoreBatchRequest" : {
        "properties" : {
          "targetDeckId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "Standalone MemoX backend API. Authentication and sync are not enabled yet.",
    "title" : "MemoX API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/cards/bulk-delete" : {
      "post" : {
        "operationId" : "deleteCards",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeleteCardsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DeleteBatchResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "One batch per card, in request order"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DeleteBatchResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Empty batch, or more than 500 ids"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DeleteBatchResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "An id names no active card"
          }
        },
        "summary" : "Move a batch of cards to Trash, opening one batch per card",
        "tags" : [
          "Trash"
        ]
      }
    },
...
    "/api/v1/decks/{deckId}" : {
      "delete" : {
        "operationId" : "deleteDeck",
        "parameters" : [
          {
            "in" : "path",
            "name" : "deckId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeleteBatchResponse"
                }
              }
            },
            "description" : "The batch the deletion opened"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeleteBatchResponse"
                }
              }
            },
            "description" : "Deck not found, or already in Trash"
          }
        },
        "summary" : "Move a deck and its whole active subtree to Trash",
        "tags" : [
          "Trash"
        ]
      },
      "get" : {
        "operationId" : "getDeck",
        "parameters" : [
          {
            "in" : "path",
            "name" : "deckId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeckResponse"
                }
              }
            },
            "description" : "Deck found"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeckResponse"
                }
              }
            },
            "description" : "Deck not found"
          }
        },
        "summary" : "Get a deck",
        "tags" : [
          "Decks"
        ]
      },
      "patch" : {
        "operationId" : "renameDeck",
        "parameters" : [
          {
            "in" : "path",
            "name" : "deckId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RenameDeckRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeckResponse"
                }
              }
            },
            "description" : "Deck renamed"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeckResponse"
                }
              }
            },
            "description" : "Invalid name"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeckResponse"
                }
              }
            },
            "description" : "Deck not found"
          }
        },
        "summary" : "Rename a deck",
        "tags" : [
          "Decks"
        ]
      }
    },
...
    "/api/v1/trash/purge" : {
      "post" : {
        "operationId" : "purge",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PurgeBatchesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PurgeReportResponse"
                }
              }
            },
            "description" : "How many batches were purged"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PurgeReportResponse"
                }
              }
            },
            "description" : "Empty selection, or more than 500 ids"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PurgeReportResponse"
                }
              }
            },
            "description" : "A named batch is no longer in Trash"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PurgeReportResponse"
                }
              }
            },
            "description" : "Mixed item types, or a cascade beyond the selection"
          }
        },
        "summary" : "Delete the named batches permanently",
        "tags" : [
          "Trash"
        ]
      }
    },
...
    "/api/v1/trash/{batchId}/restore" : {
      "post" : {
        "operationId" : "restore",
        "parameters" : [
          {
            "in" : "path",
            "name" : "batchId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestoreBatchRequest"
              }
            }
          }
        },
        "responses" : {
          "204" : {
            "description" : "The batch is back and no longer in Trash"
          },
          "404" : {
            "description" : "The batch is no longer in Trash"
          },
          "409" : {
            "description" : "The target cannot hold this item"
          }
        },
        "summary" : "Restore one batch into a chosen target",
        "tags" : [
          "Trash"
        ]
      }
    }
  }
}